Make cgab cpp compatible
