+++ test/fibers.gab
@@ -1,0 +1,1 @@
+Ranges.make(100000).each n => Fibers.make () => n
